data/temp_project/requirements.txt
streamlit
pandas
llama-cpp-python
matplotlib